# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: ThumbBoost AI API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
  - name: auth
  - name: youtube
  - name: generate
  - name: credits
  - name: history
  - name: stripe
  - name: admin
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /auth/user:
    get:
      operationId: getCurrentAuthUser
      tags: [auth]
      summary: Get current auth user
      responses:
        "200":
          description: Current user info
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GetCurrentAuthUserResponse"

  /youtube/fetch:
    post:
      operationId: fetchYouTubeVideo
      tags: [youtube]
      summary: Fetch YouTube video metadata
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/YouTubeFetchInput"
      responses:
        "200":
          description: Video metadata fetched
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/YouTubeVideoData"
        "400":
          description: Invalid URL
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /generate/all:
    post:
      operationId: generateAll
      tags: [generate]
      summary: Generate titles, description, tags, and thumbnails (costs 1 credit)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GenerateAllInput"
      responses:
        "200":
          description: All generated assets
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GenerateAllResult"
        "402":
          description: Insufficient credits
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /credits/status:
    get:
      operationId: getCreditsStatus
      tags: [credits]
      summary: Get current user credit balance and plan info
      responses:
        "200":
          description: Credit status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreditsStatus"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /history:
    get:
      operationId: getHistory
      tags: [history]
      summary: Get the current user's generation history
      responses:
        "200":
          description: Generation history
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HistoryResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /stripe/products-with-prices:
    get:
      operationId: getProductsWithPrices
      tags: [stripe]
      summary: Get all active products with their prices
      responses:
        "200":
          description: Products with prices
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductsWithPricesResponse"

  /stripe/checkout:
    post:
      operationId: createCheckout
      tags: [stripe]
      summary: Create a Stripe checkout session
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CheckoutInput"
      responses:
        "200":
          description: Checkout session URL
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CheckoutResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /stripe/portal:
    post:
      operationId: createPortalSession
      tags: [stripe]
      summary: Create a Stripe customer portal session
      responses:
        "200":
          description: Portal URL
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PortalResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /stripe/subscription:
    get:
      operationId: getSubscription
      tags: [stripe]
      summary: Get current user's subscription
      responses:
        "200":
          description: Subscription info
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SubscriptionResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /config/status:
    get:
      operationId: getConfigStatus
      tags: [generate]
      summary: Check which API keys are configured
      responses:
        "200":
          description: Config status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ConfigStatus"

  /admin/users:
    get:
      operationId: adminListUsers
      tags: [admin]
      summary: Admin — list all users
      responses:
        "200":
          description: Users list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminUsersResponse"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required: [status]

    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]

    GetCurrentAuthUserResponse:
      type: object
      properties:
        isAuthenticated:
          type: boolean
        user:
          $ref: "#/components/schemas/AuthUser"

    AuthUser:
      type: object
      properties:
        id:
          type: string
        email:
          type: ["string", "null"]
        firstName:
          type: ["string", "null"]
        lastName:
          type: ["string", "null"]
        profileImageUrl:
          type: ["string", "null"]
        plan:
          type: string
        credits:
          type: integer
        isAdmin:
          type: integer

    YouTubeFetchInput:
      type: object
      required: [url]
      properties:
        url:
          type: string

    YouTubeVideoData:
      type: object
      required: [videoId, title, description, thumbnailUrl]
      properties:
        videoId:
          type: string
        title:
          type: string
        description:
          type: string
        thumbnailUrl:
          type: string
        channelTitle:
          type: ["string", "null"]
        channelId:
          type: ["string", "null"]
        viewCount:
          type: ["string", "null"]
        publishedAt:
          type: ["string", "null"]
        channelThumbnailStyle:
          type: ["string", "null"]
          description: Summarized visual style of the channel's thumbnails for prompting

    GenerateAllInput:
      type: object
      required: [topic]
      properties:
        topic:
          type: string
        description:
          type: ["string", "null"]
        style:
          type: ["string", "null"]
        targetAudience:
          type: ["string", "null"]
        channelStyle:
          type: ["string", "null"]
          description: Channel visual style to match
        thumbnailCount:
          type: ["integer", "null"]

    ThumbnailImage:
      type: object
      required: [url, index]
      properties:
        url:
          type: string
        index:
          type: integer
        prompt:
          type: ["string", "null"]

    GenerateAllResult:
      type: object
      required: [titles, description, tags, thumbnails, generationId]
      properties:
        titles:
          type: array
          items:
            type: string
        description:
          type: string
        tags:
          type: array
          items:
            type: string
        thumbnails:
          type: array
          items:
            $ref: "#/components/schemas/ThumbnailImage"
        provider:
          type: ["string", "null"]
        generationId:
          type: string
        creditsRemaining:
          type: integer

    CreditsStatus:
      type: object
      required: [credits, plan, totalUsed]
      properties:
        credits:
          type: integer
        plan:
          type: string
        totalUsed:
          type: integer
        resetAt:
          type: ["string", "null"]

    HistoryItem:
      type: object
      required: [id, topic, createdAt]
      properties:
        id:
          type: string
        topic:
          type: string
        sourceUrl:
          type: ["string", "null"]
        titles:
          type: array
          items:
            type: string
        description:
          type: string
        tags:
          type: array
          items:
            type: string
        thumbnails:
          type: array
          items:
            $ref: "#/components/schemas/ThumbnailImage"
        provider:
          type: ["string", "null"]
        createdAt:
          type: string

    HistoryResponse:
      type: object
      required: [items]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/HistoryItem"

    ProductPrice:
      type: object
      properties:
        id:
          type: string
        unit_amount:
          type: ["integer", "null"]
        currency:
          type: string
        recurring:
          type: ["object", "null"]
        active:
          type: boolean

    Product:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        description:
          type: ["string", "null"]
        active:
          type: boolean
        prices:
          type: array
          items:
            $ref: "#/components/schemas/ProductPrice"

    ProductsWithPricesResponse:
      type: object
      required: [data]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Product"

    CheckoutInput:
      type: object
      required: [priceId]
      properties:
        priceId:
          type: string

    CheckoutResponse:
      type: object
      required: [url]
      properties:
        url:
          type: string

    PortalResponse:
      type: object
      required: [url]
      properties:
        url:
          type: string

    SubscriptionResponse:
      type: object
      properties:
        subscription:
          type: ["object", "null"]

    ConfigStatus:
      type: object
      required: [openai, youtube, replicate]
      properties:
        openai:
          type: boolean
        youtube:
          type: boolean
        replicate:
          type: boolean

    AdminUsersResponse:
      type: object
      required: [users, total]
      properties:
        users:
          type: array
          items:
            $ref: "#/components/schemas/AuthUser"
        total:
          type: integer