# lib/api-zod/src/generated/types/generateAllInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * ThumbBoost AI API
 * OpenAPI spec version: 0.1.0
 */

export interface GenerateAllInput {
  topic: string;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  style?: string | null;
  /** @nullable */
  targetAudience?: string | null;
  /**
   * Channel visual style to match
   * @nullable
   */
  channelStyle?: string | null;
  /** @nullable */
  thumbnailCount?: number | null;
}