# lib/api-zod/src/generated/types/getCurrentAuthUserResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * ThumbBoost AI API
 * OpenAPI spec version: 0.1.0
 */
import type { AuthUser } from "./authUser";

export interface GetCurrentAuthUserResponse {
  isAuthenticated?: boolean;
  user?: AuthUser;
}