# lib/api-zod/src/generated/types/product.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * ThumbBoost AI API
 * OpenAPI spec version: 0.1.0
 */
import type { ProductPrice } from "./productPrice";

export interface Product {
  id?: string;
  name?: string;
  /** @nullable */
  description?: string | null;
  active?: boolean;
  prices?: ProductPrice[];
}