# lib/api-zod/src/generated/types/productPrice.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * ThumbBoost AI API
 * OpenAPI spec version: 0.1.0
 */
import type { ProductPriceRecurring } from "./productPriceRecurring";

export interface ProductPrice {
  id?: string;
  /** @nullable */
  unit_amount?: number | null;
  currency?: string;
  /** @nullable */
  recurring?: ProductPriceRecurring;
  active?: boolean;
}